# extmod/modframebuf: Add ellipse & poly methods.