… docs/library/framebuf.rst
… examples/natmod/framebuf/framebuf.c
… py/dynruntime.h
… tests/extmod/framebuf_ellipse.py
… tests/extmod/framebuf_ellipse.py.exp
… tests/extmod/framebuf_polygon.py
… tests/extmod/framebuf_polygon.py.exp