This combines #9038 by @peterhinch and #8987 by @mbooth101 with the following changes:

- Adds a common argument handling helper for framebuf methods than handle 4+ integer arguments. This saves -80 bytes.
- Adds a optional `fill` argument to `rect` (with the aim of deprecating fill_rect). With `ellipse` and `poly` now it's wasteful to have separate `fill_ellipse` and `fill_poly` methods. This adds 32 bytes.
- Add `mp_obj_is_true` to dynruntime.h to support the fill argument. This doesn't require a mpy version jump as the table doesn't need to be updated.
- Updated the signature of `ellipse` from the original PR to match the argument order of `rect` (fill before mask).
- Adds quadrant / mask constants to the ellipse PR to clarify the way the masking works. Updated the quadrant naming in the code (Q1-Q4) to match docs.
- Made `ellipse` fill use fill_rect rather than line. This will be more efficient as they're horizontal lines, so hopefully a bit of a performance boost there (ref https://github.com/micropython/micropython/pull/9038#issuecomment-1213091428)
- All up `ellipse` is now +564 bytes.
- Made `poly` only support `array.array`. This saves about 150 bytes.
- Other improvements to poly: replaced `float` calculation with integer, removed heap allocation, minor size optimisations. All up `poly` is now +616 bytes (the original PR was just over 1k).

(All size diffs calculated on PYBV11)

Note that only `ellipse` is added to the natmod version of framebuf as dynruntime.h doesn't support `mp_binary_get_size` & `mp_binary_get_val_array` required by `poly`. We can do that in the next mpy version jump.

Thank you @peterhinch and @mbooth101 for the work on this, it's a great improvement to framebuf!