@mbooth101 2022-08-12
You re-wrote my PR so that I can no longer pass the polygon coordinates a list of lists/list of tuples :-1: 

I'm trying to be very responsive feedback on my original PR, why not discuss changes there?