@peterhinch 2022-08-12
Ellipse goes like a bat out of hell now. Using the same script and test conditions, times are
 * Line draw 306μs
 * Filled 666μs (!)