@jimmo 2022-08-12
Just a quick note on the efficiency of the `array.array` approach, further to what @dpgeorge said in #8987.
Because this uses `mp_get_buffer`, it also works with anything else that supports the buffer protocol. So for small polygons (coordinates less than 256), you can use `bytes`, and in particular a literal bytes string, which means you get to use a const string and have no copies at all. If your code is frozen, it means your polygon data will be drawn from ROM.

So for example, the "M" polygon from the test cases could be written as:

```
poly = const(b'\x00\x14\x03\x14\x03\n\x06\x11\t\n\t\x14\x0c\x14\x0c\x03\t\x03\x06\n\x03\x03\x00\x03')
```

Obviously this isn't great for readability, but if your program has a number of polygons then these can be generated programatically.

> Ellipse goes like a bat out of hell now.

Thanks for testing @peterhinch !