@mbooth101 2022-08-12
> > You re-wrote my PR so that I can no longer pass the polygon coordinates a list of lists/list of tuples -1
> > I'm trying to be very responsive feedback on my original PR, why not discuss changes there?
> 
> I'm just trying to do what I can do as a maintainer to get your code merged as quickly and easily as possible without making you duplicate the effort that I just spent my Friday evening on. Not trying to diminish your work or anything and I'm sorry that it seems that way! I'm glad to see these features added to framebuf.
> 
> These two features represent an extremely significant code size increase (+1132 bytes total) and merging this is not an easy decision. I spent a good chunk of time discussing this with Damien today and then more time trying to optimise the two PRs. Sure I can turn all that work into a bunch of comments on your PR and then go backwards-and-forwards on it for the next days/weeks but I'd still need to make a PR for the first two commits, and now we'd have three PRs to wrangle.
> 
> I definitely agree that list-of-lists or lists-of-tuples is syntactically convenient, but I don't feel that it's worth either the code size or the runtime RAM cost.

Thanks -- this is all important context that I was missing!

As an outside contributor I am asked to sign off on submissions, so it seemed against the spirit of that policy to merge something *different* to what I signed off on without at least some explanation of what was going on. :-)