@jimmo 2022-08-12
> As an outside contributor I am asked to sign off on submissions, so it seemed against the spirit of that policy to merge something different to what I signed off on without at least some explanation of what was going on. :-)

Yes this is a very good point. I just left the commits with the original author (of course) but didn't think about the sign off. Our main intention here is that the sign off is conveying that the contribution meets the licence requirements and the code doesn't have any encumbrances.

This is actually something we do all the time (make changes to commits during merge while preserving the author/signoff), although usually the changes are minor or squashes/rebases. @dpgeorge we should probably clarify this particular scenario.