@peterhinch 2022-08-13
@jimmo I now have prototype rounded rectangle (filled and outline) C code. Naturally this is dependent on `.ellipse`: I'll submit a PR once this one is merged.

The idea of shapes in ROM is excellent. On a typical display all shapes are likely to be small enough to fit in a `bytes` instance. Writers of traditional arcade games should have a field day.

I'm wondering if there's scope for a [font_to_py.py](https://github.com/peterhinch/micropython-font-to-py) type utility for shapes. This would convert a set of shapes (somehow defined) into Python source that could be frozen as bytecode. The Python source would encode a data structure and an access method. Invoking the access method with an index would return a memory view into the `bytes` object of the chosen shape.

Like `font_to_py.py` it would mean that shapes could be defined on a PC in a user friendly way and could be changed without altering the application source. They would act like a resource in a conventional GUI.

I can't see any particular technical difficulty but is there likely to be any interest, I wonder?