@mbooth101 2022-08-13
Hi @jimmo this morning I tested this series of patches on my ESP32 device and have a couple of comments.

Because I have to take the extra steps of flattening my point lists and converting it to an `array.array()` I lose about 10ms per frame when rendering a scene of 104 triangles -- this no big deal, I can probably claw back some performance by refactoring my application to avoid this extra work.

Much more concerning to me are the visual defects that have been introduced:

https://user-images.githubusercontent.com/597661/184479190-920f3ada-0fcc-4469-930a-9faa05635f04.mp4

Compare with the behaviour of my original PR:

https://user-images.githubusercontent.com/597661/182123844-fca92232-0b1e-4d91-830c-54877b67a679.mp4

I will try to come up with a small reproducing test case.