@mbooth101 2022-08-13
> @mbooth101 Your videos aren't visible here.

Works fine for me. It's just an MP4 file in a HTML5 <video> tag, so maybe your browser doesn't know how to decode MP4?

In any case you can download the videos for offline viewing by clicking these links:

- New broken behaviour: [clicky clicky](https://user-images.githubusercontent.com/597661/184479190-920f3ada-0fcc-4469-930a-9faa05635f04.mp4)
- Old correct behaviour: [clicky clicky](https://user-images.githubusercontent.com/597661/182123844-fca92232-0b1e-4d91-830c-54877b67a679.mp4)