@mbooth101 2022-08-13
@jimmo Here's a small repoducer for the issue I see in the above video:

```
import sys
import framebuf
from array import array

def print_buffer(buffer, width, height):
    for row in range(height):
        for col in range(width):
            val = buffer[(row * width) + col]
            sys.stdout.write(" {:02x}".format(val) if val else " ··")
        sys.stdout.write("\n")

t1 = array('h', [40, 0, 20, 68, 62, 40])
t2 = array('h', [40, 0, 0, 16, 20, 68])

w = 70
h = 70
buf = bytearray(w * h)
fbuf = framebuf.FrameBuffer(buf, w, h, framebuf.GS8)

fbuf.poly(0,0,t1,0xff, True)
fbuf.poly(0,0,t2,0xff, True)

print_buffer(buf, w, h)
```