@jimmo 2022-08-15
@mbooth101 Based on the idea that drawing the outline can solve this, I have implemented a more optimised fix for this. The idea is to detect the "lost" pixels during the node calculations, but rather than getting the node list to account for these, just explicitly draw the pixels there (either via setpixel or line).

This adds +72 bytes on PYBV11, but makes filled poly() about 10-15% faster.

Baseline is 32 ms/frame (empty scene no polygons)
- Test 1 with 40 triangles goes from 53 ms/frame to 50 ms/frame (21 ms --> 18 ms for poly)
- Test 2 with 20 stars goes from 49 ms/frame to 47 ms/frame (17 --> 15 ms for poly)

I've put it in a separate commit... we can squash during merge if desired.

It's still not exactly perfect -- there are some pixels along diagonal edges that it misses, but visually it's pretty good.